JIRA-0000:[Fix] woorked on the data struture on find patient endpoint

+++ src/main/java/org/lamisplus/modules/pmtct/domain/entity/Infant.java
@@ -29,7 +29,7 @@ public class Infant extends PMTCTTransactionalEntity implements Serializable, Pe
     private Double bodyWeight;
     private String ctxStatus;
     private Long pmtctCycleId;
-    private Integer archived;
+    private Long archived;
     private String source;
 
     @PrePersist
